[V4] Sketch of promise structure (#53)
* explicit casts

* add concpets and task

* primatives in promise

* instanciate

* use coroutine fib

* golf

* split up macros

* use split macros

* rename

* move common to fib

* tweak clang tidy

* impl terminate_with

* drop unused

* break out frame

* utiliy types

* formatting

* wip

* utility tests

* task immovable

* rename test

* drop test form bench preset

* add sym transfer

* add more assume

* disable false +ve lint

* movable task

* golf

* assume

* alloc policy structure

* more alloc policy

* inherit from policy

* add dummy

* initial test

* constants

* sized delete

* renames

* bump

* use thread local

* new assert

* fix assume

* protect macro

* fix macro typo

* typo

* rm bad comment

* add to file set

+++ benchmark/src/libfork_benchmark/fib/serial_return.cpp
@@ -1,6 +1,7 @@
 #include <benchmark/benchmark.h>
 
-#include "libfork_benchmark/common.hpp"
+#include "libfork/__impl/compiler.hpp"
+
 #include "libfork_benchmark/fib/fib.hpp"
 
 namespace {
@@ -21,7 +22,7 @@ void fib_serial_return(benchmark::State &state) {
   std::int64_t n = state.range(0);
   std::int64_t expect = fib_ref(n);
 
-  state.counters["n"] = n;
+  state.counters["n"] = static_cast<double>(n);
 
   for (auto _ : state) {
     benchmark::DoNotOptimize(n);
